Search for data breaches associated with one or more accounts.
Source: R/account_breaches.R

Search for data breaches associated with one or more accounts.

account_breaches(accounts, truncate = FALSE, domain = NULL,
  as_list = TRUE, ...)

Arguments

accounts
A character vector of email addresses or account names

truncate
Whether responses should be trimmed to just name of breached site

domain
Search a specific domain

as_list
whether to return results as a list (default, TRUE) or a data.frame

...
Optional passthrough to HIBP_headers()

Value

List of data.frames containing results

Examples

account_breaches(c("steff.sullivan@gmail.com","stephanie.g.locke@gmail.com"))

#> $`steff.sullivan@gmail.com`
#>             Title           Name       Domain BreachDate            AddedDate
#> 1           Adobe          Adobe    adobe.com 2013-10-04 2013-12-04T00:00:00Z
#> 2          Disqus         Disqus   disqus.com 2012-07-01 2017-10-06T23:03:51Z
#> 3        LinkedIn       LinkedIn linkedin.com 2012-05-05 2016-05-21T21:35:40Z
#> 4 Onliner Spambot OnlinerSpambot              2017-08-28 2017-08-29T19:25:56Z
#>           ModifiedDate  PwnCount
#> 1 2013-12-04T00:00:00Z 152445165
#> 2 2017-10-06T23:03:51Z  17551044
#> 3 2016-05-21T21:35:40Z 164611595
#> 4 2017-08-29T19:25:56Z 711477622
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        Description
#> 1                                                                                                                                                                                  In October 2013, 153 million Adobe accounts were breached with each containing an internal ID, username, email, <em>encrypted</em> password and a password hint in plain text. The password cryptography was poorly done and <a href="http://stricture-group.com/files/adobe-top100.txt" target="_blank" rel="noopener">many were quickly resolved back to plain text</a>. The unencrypted hints also <a href="http://www.troyhunt.com/2013/11/adobe-credentials-and-serious.html" target="_blank" rel="noopener">disclosed much about the passwords</a> adding further to the risk that hundreds of millions of Adobe customers already faced.
#> 2                                                                                                                                                                                                                                                                          In October 2017, the blog commenting service <a href="https://blog.disqus.com/security-alert-user-info-breach" target="_blank" rel="noopener">Disqus announced they'd suffered a data breach</a>. The breach dated back to July 2012 but wasn't identified until years later when the data finally surfaced. The breach contained over 17.5 million unique email addresses and usernames. Users who created logins on Disqus had salted SHA1 hashes of passwords whilst users who logged in via social providers only had references to those accounts.
#> 3                                                                                                                                                                                                                                                                                                                         In May 2016, <a href="https://www.troyhunt.com/observations-and-thoughts-on-the-linkedin-data-breach" target="_blank" rel="noopener">LinkedIn had 164 million email addresses and passwords exposed</a>. Originally hacked in 2012, the data remained out of sight until being offered for sale on a dark market site 4 years later. The passwords in the breach were stored as SHA1 hashes without salt, the vast majority of which were quickly cracked in the days following the release of the data.
#> 4 In August 2017, a spambot by the name of <a href="https://benkowlab.blogspot.com.au/2017/08/from-onliner-spambot-to-millions-of.html" target="_blank" rel="noopener">Onliner Spambot was identified by security researcher Benkow mo<U+029E>u<U+018E>q</a>. The malicious software contained a server-based component located on an IP address in the Netherlands which exposed a large number of files containing personal information. In total, there were 711 million unique email addresses, many of which were also accompanied by corresponding passwords. A full write-up on what data was found is in the blog post titled <a href="https://www.troyhunt.com/inside-the-massive-711-million-record-onliner-spambot-dump" target="_blank" rel="noopener">Inside the Massive 711 Million Record Onliner Spambot Dump</a>.
#>                                             DataClasses IsVerified IsFabricated
#> 1 Email addresses, Password hints, Passwords, Usernames       TRUE        FALSE
#> 2                 Email addresses, Passwords, Usernames       TRUE        FALSE
#> 3                            Email addresses, Passwords       TRUE        FALSE
#> 4                            Email addresses, Passwords       TRUE        FALSE
#>   IsSensitive IsActive IsRetired IsSpamList LogoType
#> 1       FALSE     TRUE     FALSE      FALSE      svg
#> 2       FALSE     TRUE     FALSE      FALSE      svg
#> 3       FALSE     TRUE     FALSE      FALSE      svg
#> 4       FALSE     TRUE     FALSE       TRUE      png
#> 
#> $`stephanie.g.locke@gmail.com`
#>      Title     Name       Domain BreachDate            AddedDate
#> 1    Bitly    Bitly    bitly.com 2014-05-08 2017-10-06T06:31:50Z
#> 2 GeekedIn GeekedIn geekedin.net 2016-08-15 2016-11-17T19:44:24Z
#>           ModifiedDate PwnCount
#> 1 2017-10-06T08:05:10Z  9313136
#> 2 2016-11-17T19:44:24Z  1073164
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           Description
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                    In May 2014, the link management company <a href="https://bitly.com/blog/urgent-security-update-regarding-your-bitly-account/" target="_blank" rel="noopener">Bitly announced they'd suffered a data breach</a>. The breach contained over 9.3 million unique email addresses, usernames and hashed passwords, most using SHA1 with a small number using bcrypt.
#> 2 In August 2016, the technology recruitment site <a href="http://geekedin.net" target="_blank" rel="noopener">GeekedIn</a> left a MongoDB database exposed and over 8M records were extracted by an unknown third party. The breached data was originally scraped from GitHub in violation of their terms of use and contained information exposed in public profiles, including over 1 million members' email addresses. Full details on the incident (including how impacted members can see their leaked data) are covered in the blog post on <a href="https://www.troyhunt.com/8-million-github-profiles-were-leaked-from-geekedins-mongodb-heres-how-to-see-yours" target="_blank" rel="noopener">8 million GitHub profiles were leaked from GeekedIn's MongoDB - here's how to see yours</a>.
#>                                                                                                      DataClasses
#> 1                                                                          Email addresses, Passwords, Usernames
#> 2 Email addresses, Geographic locations, Names, Professional skills, Usernames, Years of professional experience
#>   IsVerified IsFabricated IsSensitive IsActive IsRetired IsSpamList LogoType
#> 1       TRUE        FALSE       FALSE     TRUE     FALSE      FALSE      svg
#> 2       TRUE        FALSE       FALSE     TRUE     FALSE      FALSE      png
#>