Get a specific breached site, based in breach name (not domain)
Source: R/breached_site.R

Get a specific breached site, based in breach name (not domain)

breached_site(name, verbose = TRUE, agent = NULL)

Arguments

name
Search for a specific named breach

verbose
whether to message about http errors and re-tries

agent
agent to be used as header in calls, by default "HIBPwned R pkg". # nolint

Value

Data.frame containing breach details

Details

Note that the package uses memoise (https://github.com/r-lib/memoise) with no timeout argument so that results are cached inside an active R session.

Examples

breached_site("Adobe")

#>   Title  Name    Domain BreachDate            AddedDate         ModifiedDate
#> 1 Adobe Adobe adobe.com 2013-10-04 2013-12-04T00:00:00Z 2013-12-04T00:00:00Z
#>    PwnCount
#> 1 152445165
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       Description
#> 1 In October 2013, 153 million Adobe accounts were breached with each containing an internal ID, username, email, <em>encrypted</em> password and a password hint in plain text. The password cryptography was poorly done and <a href="http://stricture-group.com/files/adobe-top100.txt" target="_blank" rel="noopener">many were quickly resolved back to plain text</a>. The unencrypted hints also <a href="http://www.troyhunt.com/2013/11/adobe-credentials-and-serious.html" target="_blank" rel="noopener">disclosed much about the passwords</a> adding further to the risk that hundreds of millions of Adobe customers already faced.
#>                                             DataClasses IsVerified IsFabricated
#> 1 Email addresses, Password hints, Passwords, Usernames       TRUE        FALSE
#>   IsSensitive IsActive IsRetired IsSpamList LogoType
#> 1       FALSE     TRUE     FALSE      FALSE      svg