Get list of classes of data that have been exposed in breaches

Get list of classes of data that have been exposed in breaches

data_classes(...)

Arguments

...
Optional passthrough to HIBP_headers()

Value

Data.frame containing data class details

Examples

data_classes()

#>  [1] "Account balances"                 "Age groups"                      
#>  [3] "Astrological signs"               "Auth tokens"                     
#>  [5] "Avatars"                          "Bank account numbers"            
#>  [7] "Banking PINs"                     "Beauty ratings"                  
#>  [9] "Biometric data"                   "Browser user agent details"      
#> [11] "Buying preferences"               "Car ownership statuses"          
#> [13] "Career levels"                    "Charitable donations"            
#> [15] "Chat logs"                        "Credit card CVV"                 
#> [17] "Credit cards"                     "Credit status information"       
#> [19] "Customer feedback"                "Customer interactions"           
#> [21] "Dates of birth"                   "Deceased date"                   
#> [23] "Device information"               "Device usage tracking data"      
#> [25] "Drinking habits"                  "Drug habits"                     
#> [27] "Education levels"                 "Email addresses"                 
#> [29] "Email messages"                   "Employers"                       
#> [31] "Ethnicities"                      "Family members' names"           
#> [33] "Family plans"                     "Family structure"                
#> [35] "Financial investments"            "Financial transactions"          
#> [37] "Fitness levels"                   "Genders"                         
#> [39] "Geographic locations"             "Government issued IDs"           
#> [41] "Health insurance information"     "Historical passwords"            
#> [43] "Home ownership statuses"          "Homepage URLs"                   
#> [45] "Income levels"                    "Instant messenger identities"    
#> [47] "IP addresses"                     "Job titles"                      
#> [49] "MAC addresses"                    "Marital statuses"                
#> [51] "Names"                            "Net worths"                      
#> [53] "Nicknames"                        "Parenting plans"                 
#> [55] "Partial credit card data"         "Passport numbers"                
#> [57] "Password hints"                   "Passwords"                       
#> [59] "Payment histories"                "Payment methods"                 
#> [61] "Personal descriptions"            "Personal health data"            
#> [63] "Personal interests"               "Phone numbers"                   
#> [65] "Physical addresses"               "Physical attributes"             
#> [67] "Political donations"              "Political views"                 
#> [69] "Private messages"                 "Professional skills"             
#> [71] "Purchases"                        "Purchasing habits"               
#> [73] "Races"                            "Recovery email addresses"        
#> [75] "Relationship statuses"            "Religions"                       
#> [77] "Reward program balances"          "Salutations"                     
#> [79] "Security questions and answers"   "Sexual fetishes"                 
#> [81] "Sexual orientations"              "Smoking habits"                  
#> [83] "SMS messages"                     "Social connections"              
#> [85] "Spoken languages"                 "Survey results"                  
#> [87] "Time zones"                       "Travel habits"                   
#> [89] "User statuses"                    "User website URLs"               
#> [91] "Usernames"                        "Utility bills"                   
#> [93] "Vehicle details"                  "Website activity"                
#> [95] "Work habits"                      "Years of birth"                  
#> [97] "Years of professional experience"