Search for data breaches associated with one or more accounts.

Search for data breaches associated with one or more accounts.

account_breaches(accounts, truncate = FALSE, domain = NULL,
  as_list = TRUE, verbose = TRUE, agent = NULL)

Arguments

accounts
A character vector of email addresses or account names

truncate
Whether responses should be trimmed to just name of breached site

domain
Search a specific domain

as_list
whether to return results as a list (default, TRUE) or a data.frame, When returned as a data.frame only accounts with at least one breach are returned.

verbose
whether to message about http errors and re-tries

agent
agent to be used as header in calls, by default "HIBPwned R pkg". # nolint

Value

List of data.frames containing results

Details

Note that the package uses memoise (https://github.com/r-lib/memoise) with no timeout argument so that results are cached inside an active R session.

Examples

account_breaches(c("steff.sullivan@gmail.com","stephanie.g.locke@gmail.com"))

#> http error code: 403

#> Try number 2

#> http error code: 403

#> Try number 3

#> http error code: 403

#> Try number 4

#> http error code: 403

#> Try number 5

#> http error code: 403

#> Try number 6

#> This is the last try, if it fails will return NULL

#> http error code: 403

#> http error code: 403

#> Try number 2

#> http error code: 403

#> Try number 3

#> http error code: 403

#> Try number 4

#> http error code: 403

#> Try number 5

#> http error code: 403

#> Try number 6

#> This is the last try, if it fails will return NULL

#> http error code: 403

#> $`steff.sullivan@gmail.com`
#> NULL
#> 
#> $`stephanie.g.locke@gmail.com`
#> NULL
#>