Search for data breaches associated with one or more accounts.

Search for data breaches associated with one or more accounts.

account_breaches(accounts, truncate = FALSE, domain = NULL, ...)

Arguments

accounts
A character vector of email addresses or account names

truncate
Whether responses should be trimmed to just name of breached site

domain
Search a specific domain

...
Optional passthrough to HIBP_headers()

Value

List of data.frames containing results

Examples

account_breaches(c("steff.sullivan@gmail.com","stephanie.g.locke@gmail.com"))

#> $`steff.sullivan@gmail.com`
#>      Title     Name       Domain BreachDate            AddedDate
#> 1    Adobe    Adobe    adobe.com 2013-10-04 2013-12-04T00:00:00Z
#> 2 LinkedIn LinkedIn linkedin.com 2012-05-05 2016-05-21T21:35:40Z
#>           ModifiedDate  PwnCount
#> 1 2013-12-04T00:00:00Z 152445165
#> 2 2016-05-21T21:35:40Z 164611595
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       Description
#> 1 In October 2013, 153 million Adobe accounts were breached with each containing an internal ID, username, email, <em>encrypted</em> password and a password hint in plain text. The password cryptography was poorly done and <a href="http://stricture-group.com/files/adobe-top100.txt" target="_blank" rel="noopener">many were quickly resolved back to plain text</a>. The unencrypted hints also <a href="http://www.troyhunt.com/2013/11/adobe-credentials-and-serious.html" target="_blank" rel="noopener">disclosed much about the passwords</a> adding further to the risk that hundreds of millions of Adobe customers already faced.
#> 2                                                                                                                                        In May 2016, <a href="https://www.troyhunt.com/observations-and-thoughts-on-the-linkedin-data-breach" target="_blank" rel="noopener">LinkedIn had 164 million email addresses and passwords exposed</a>. Originally hacked in 2012, the data remained out of sight until being offered for sale on a dark market site 4 years later. The passwords in the breach were stored as SHA1 hashes without salt, the vast majority of which were quickly cracked in the days following the release of the data.
#>                                             DataClasses IsVerified IsFabricated
#> 1 Email addresses, Password hints, Passwords, Usernames       TRUE        FALSE
#> 2                            Email addresses, Passwords       TRUE        FALSE
#>   IsSensitive IsActive IsRetired IsSpamList LogoType
#> 1       FALSE     TRUE     FALSE      FALSE      svg
#> 2       FALSE     TRUE     FALSE      FALSE      svg
#> 
#> $`stephanie.g.locke@gmail.com`
#>      Title     Name       Domain BreachDate            AddedDate
#> 1 GeekedIn GeekedIn geekedin.net 2016-08-15 2016-11-17T19:44:24Z
#>           ModifiedDate PwnCount
#> 1 2016-11-17T19:44:24Z  1073164
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           Description
#> 1 In August 2016, the technology recruitment site <a href="http://geekedin.net" target="_blank" rel="noopener">GeekedIn</a> left a MongoDB database exposed and over 8M records were extracted by an unknown third party. The breached data was originally scraped from GitHub in violation of their terms of use and contained information exposed in public profiles, including over 1 million members' email addresses. Full details on the incident (including how impacted members can see their leaked data) are covered in the blog post on <a href="https://www.troyhunt.com/8-million-github-profiles-were-leaked-from-geekedins-mongodb-heres-how-to-see-yours" target="_blank" rel="noopener">8 million GitHub profiles were leaked from GeekedIn's MongoDB - here's how to see yours</a>.
#>                                                                                                      DataClasses
#> 1 Email addresses, Geographic locations, Names, Professional skills, Usernames, Years of professional experience
#>   IsVerified IsFabricated IsSensitive IsActive IsRetired IsSpamList LogoType
#> 1       TRUE        FALSE       FALSE     TRUE     FALSE      FALSE      png
#>