Get list of classes of data that have been exposed in breaches
Source: R/data_classes.R

Get list of classes of data that have been exposed in breaches

data_classes(verbose = TRUE, agent = NULL)

Arguments

verbose
whether to message about http errors and re-tries

agent
agent to be used as header in calls, by default "HIBPwned R pkg". # nolint

Value

Data.frame containing data class details

Details

Note that the package uses memoise (https://github.com/r-lib/memoise) with no timeout argument so that results are cached inside an active R session.

Examples

data_classes()

#>   [1] "Account balances"                 "Address book contacts"           
#>   [3] "Age groups"                       "Ages"                            
#>   [5] "Apps installed on devices"        "Astrological signs"              
#>   [7] "Auth tokens"                      "Avatars"                         
#>   [9] "Bank account numbers"             "Banking PINs"                    
#>  [11] "Beauty ratings"                   "Biometric data"                  
#>  [13] "Browser user agent details"       "Buying preferences"              
#>  [15] "Car ownership statuses"           "Career levels"                   
#>  [17] "Cellular network names"           "Charitable donations"            
#>  [19] "Chat logs"                        "Credit card CVV"                 
#>  [21] "Credit cards"                     "Credit status information"       
#>  [23] "Customer feedback"                "Customer interactions"           
#>  [25] "Dates of birth"                   "Deceased date"                   
#>  [27] "Deceased statuses"                "Device information"              
#>  [29] "Device usage tracking data"       "Drinking habits"                 
#>  [31] "Drug habits"                      "Eating habits"                   
#>  [33] "Education levels"                 "Email addresses"                 
#>  [35] "Email messages"                   "Employers"                       
#>  [37] "Ethnicities"                      "Family members' names"           
#>  [39] "Family plans"                     "Family structure"                
#>  [41] "Financial investments"            "Financial transactions"          
#>  [43] "Fitness levels"                   "Genders"                         
#>  [45] "Geographic locations"             "Government issued IDs"           
#>  [47] "Health insurance information"     "Historical passwords"            
#>  [49] "Home ownership statuses"          "Homepage URLs"                   
#>  [51] "IMEI numbers"                     "IMSI numbers"                    
#>  [53] "Income levels"                    "Instant messenger identities"    
#>  [55] "IP addresses"                     "Job titles"                      
#>  [57] "MAC addresses"                    "Marital statuses"                
#>  [59] "Names"                            "Nationalities"                   
#>  [61] "Net worths"                       "Nicknames"                       
#>  [63] "Parenting plans"                  "Partial credit card data"        
#>  [65] "Passport numbers"                 "Password hints"                  
#>  [67] "Passwords"                        "Payment histories"               
#>  [69] "Payment methods"                  "Personal descriptions"           
#>  [71] "Personal health data"             "Personal interests"              
#>  [73] "Phone numbers"                    "Physical addresses"              
#>  [75] "Physical attributes"              "Political donations"             
#>  [77] "Political views"                  "Private messages"                
#>  [79] "Professional skills"              "Profile photos"                  
#>  [81] "Purchases"                        "Purchasing habits"               
#>  [83] "Races"                            "Recovery email addresses"        
#>  [85] "Relationship statuses"            "Religions"                       
#>  [87] "Reward program balances"          "Salutations"                     
#>  [89] "School grades (class levels)"     "Security questions and answers"  
#>  [91] "Sexual fetishes"                  "Sexual orientations"             
#>  [93] "Smoking habits"                   "SMS messages"                    
#>  [95] "Social connections"               "Social media profiles"           
#>  [97] "Spoken languages"                 "Support tickets"                 
#>  [99] "Survey results"                   "Time zones"                      
#> [101] "Travel habits"                    "User statuses"                   
#> [103] "User website URLs"                "Usernames"                       
#> [105] "Utility bills"                    "Vehicle details"                 
#> [107] "Website activity"                 "Work habits"                     
#> [109] "Years of birth"                   "Years of professional experience"