Get list of classes of data that have been exposed in breaches

Get list of classes of data that have been exposed in breaches

data_classes(verbose = TRUE, agent = NULL)

Arguments

verbose
whether to message about http errors and re-tries

agent
agent to be used as header in calls, by default "HIBPwned R pkg". # nolint

Value

Data.frame containing data class details

Details

Note that the package uses memoise (https://github.com/r-lib/memoise) with no timeout argument so that results are cached inside an active R session.

Examples

data_classes()

#>   [1] "Account balances"                 "Address book contacts"           
#>   [3] "Age groups"                       "Ages"                            
#>   [5] "Apps installed on devices"        "Astrological signs"              
#>   [7] "Auth tokens"                      "Avatars"                         
#>   [9] "Bank account numbers"             "Beauty ratings"                  
#>  [11] "Biometric data"                   "Browser user agent details"      
#>  [13] "Buying preferences"               "Car ownership statuses"          
#>  [15] "Career levels"                    "Cellular network names"          
#>  [17] "Charitable donations"             "Chat logs"                       
#>  [19] "Credit card CVV"                  "Credit cards"                    
#>  [21] "Credit status information"        "Customer feedback"               
#>  [23] "Customer interactions"            "Dates of birth"                  
#>  [25] "Deceased date"                    "Deceased statuses"               
#>  [27] "Device information"               "Device usage tracking data"      
#>  [29] "Drinking habits"                  "Drug habits"                     
#>  [31] "Eating habits"                    "Education levels"                
#>  [33] "Email addresses"                  "Email messages"                  
#>  [35] "Employers"                        "Ethnicities"                     
#>  [37] "Family members' names"            "Family plans"                    
#>  [39] "Family structure"                 "Financial investments"           
#>  [41] "Financial transactions"           "Fitness levels"                  
#>  [43] "Genders"                          "Geographic locations"            
#>  [45] "Government issued IDs"            "Health insurance information"    
#>  [47] "Historical passwords"             "Home loan information"           
#>  [49] "Home ownership statuses"          "Homepage URLs"                   
#>  [51] "IMEI numbers"                     "IMSI numbers"                    
#>  [53] "Income levels"                    "Instant messenger identities"    
#>  [55] "IP addresses"                     "Job titles"                      
#>  [57] "MAC addresses"                    "Marital statuses"                
#>  [59] "Names"                            "Nationalities"                   
#>  [61] "Net worths"                       "Nicknames"                       
#>  [63] "Occupations"                      "Parenting plans"                 
#>  [65] "Partial credit card data"         "Passport numbers"                
#>  [67] "Password hints"                   "Passwords"                       
#>  [69] "Payment histories"                "Payment methods"                 
#>  [71] "Personal descriptions"            "Personal health data"            
#>  [73] "Personal interests"               "Phone numbers"                   
#>  [75] "Physical addresses"               "Physical attributes"             
#>  [77] "PINs"                             "Political donations"             
#>  [79] "Political views"                  "Private messages"                
#>  [81] "Professional skills"              "Profile photos"                  
#>  [83] "Purchases"                        "Purchasing habits"               
#>  [85] "Races"                            "Recovery email addresses"        
#>  [87] "Relationship statuses"            "Religions"                       
#>  [89] "Reward program balances"          "Salutations"                     
#>  [91] "School grades (class levels)"     "Security questions and answers"  
#>  [93] "Sexual fetishes"                  "Sexual orientations"             
#>  [95] "Smoking habits"                   "SMS messages"                    
#>  [97] "Social connections"               "Social media profiles"           
#>  [99] "Spoken languages"                 "Support tickets"                 
#> [101] "Survey results"                   "Time zones"                      
#> [103] "Travel habits"                    "User statuses"                   
#> [105] "User website URLs"                "Usernames"                       
#> [107] "Utility bills"                    "Vehicle details"                 
#> [109] "Website activity"                 "Work habits"                     
#> [111] "Years of birth"                   "Years of professional experience"